Add first html files for the recipe project
Add index.html as a home page for the project, and add
chicken-wings.html as my first recipe in the website. They are just
placeholder files for now with the usual HTML boilerplate code added to
them. Also updated the README a little bit to include some basic
information, nothing set in stone yet.